requirements.txt
pyyaml>=6.0
websocket-client>=1.6.0
requests>=2.31.0
Pillow>=10.0.0
